requirements.txt
boto3==1.34.127
Django==5.0.6
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
pytest==8.2.2
ruff==0.4.9